qcacld-3.0: Mem Zero only the keys and not the key_req
Currently the driver does mem zero of the entire structure of
tLimMlmSetKeysReq in API lim_process_mlm_set_keys_req, and
hence the STA's peer mac address also becomes zero, and
hence at further time, the driver is not able to find the
STA ID of peer mac address ( which was set as zero ), thus
when DHCP requests comes, the driver rejects it as the STA
was not in authenticated mode.

Fix is to just do mem zero of the keys, and make num of keys
as zero.

Change-Id: I7f26edfa1be9a6707c90932d0a49c963f59336c5
CRs-Fixed: 2404704